+++ misskey-api/src/endpoint/antennas/create.rs
@@ -3,6 +3,7 @@ use crate::model::user_group::UserGroup;
 use crate::model::{
     antenna::{Antenna, AntennaSource},
     id::Id,
+    query::Query,
     user_list::UserList,
 };
 
@@ -23,10 +24,12 @@ pub struct Request {
     #[cfg_attr(docsrs, doc(cfg(feature = "12-10-0")))]
     #[builder(default, setter(strip_option))]
     pub user_group_id: Option<Id<UserGroup>>,
-    pub keywords: Vec<Vec<String>>,
+    #[builder(default, setter(into))]
+    pub keywords: Query<String>,
     #[cfg(feature = "12-19-0")]
     #[cfg_attr(docsrs, doc(cfg(feature = "12-19-0")))]
-    pub exclude_keywords: Vec<Vec<String>>,
+    #[builder(default, setter(into))]
+    pub exclude_keywords: Query<String>,
     pub users: Vec<String>,
     pub case_sensitive: bool,
     pub with_replies: bool,
@@ -46,7 +49,7 @@ mod tests {
 
     #[tokio::test]
     async fn request_all() {
-        use crate::model::antenna::AntennaSource;
+        use crate::model::{antenna::AntennaSource, query::Query};
 
         let client = TestClient::new();
         client
@@ -57,9 +60,9 @@ mod tests {
                 user_list_id: None,
                 #[cfg(feature = "12-10-0")]
                 user_group_id: None,
-                keywords: vec![vec!["hello".to_string(), "awesome".to_string()]],
+                keywords: Query::from_vec(vec![vec!["hello".to_string(), "awesome".to_string()]]),
                 #[cfg(feature = "12-19-0")]
-                exclude_keywords: vec![vec!["no".to_string()]],
+                exclude_keywords: Query::from_vec(vec![vec!["no".to_string()]]),
                 users: Vec::new(),
                 case_sensitive: false,
                 with_replies: false,
@@ -71,7 +74,7 @@ mod tests {
 
     #[tokio::test]
     async fn request_home() {
-        use crate::model::antenna::AntennaSource;
+        use crate::model::{antenna::AntennaSource, query::Query};
 
         let client = TestClient::new();
         client
@@ -81,9 +84,9 @@ mod tests {
                 user_list_id: None,
                 #[cfg(feature = "12-10-0")]
                 user_group_id: None,
-                keywords: vec![vec!["hey".to_string()], vec!["wow".to_string()]],
+                keywords: Query::from_vec(vec![vec!["hey".to_string()], vec!["wow".to_string()]]),
                 #[cfg(feature = "12-19-0")]
-                exclude_keywords: Vec::new(),
+                exclude_keywords: Query::default(),
                 users: Vec::new(),
                 case_sensitive: true,
                 with_replies: false,
@@ -95,7 +98,7 @@ mod tests {
 
     #[tokio::test]
     async fn request_list() {
-        use crate::model::antenna::AntennaSource;
+        use crate::model::{antenna::AntennaSource, query::Query};
 
         let client = TestClient::new();
         let list = client
@@ -111,9 +114,15 @@ mod tests {
                 user_list_id: Some(list.id),
                 #[cfg(feature = "12-10-0")]
                 user_group_id: None,
-                keywords: vec![vec!["kawaii".to_string()], vec!["cute".to_string()]],
+                keywords: Query::from_vec(vec![
+                    vec!["kawaii".to_string()],
+                    vec!["cute".to_string()],
+                ]),
                 #[cfg(feature = "12-19-0")]
-                exclude_keywords: vec![vec!["sensitive".to_string()], vec!["violence".to_string()]],
+                exclude_keywords: Query::from_vec(vec![
+                    vec!["sensitive".to_string()],
+                    vec!["violence".to_string()],
+                ]),
                 users: Vec::new(),
                 case_sensitive: false,
                 with_replies: true,
@@ -126,7 +135,7 @@ mod tests {
     #[tokio::test]
     #[cfg(feature = "12-10-0")]
     async fn request_group() {
-        use crate::model::antenna::AntennaSource;
+        use crate::model::{antenna::AntennaSource, query::Query};
 
         let client = TestClient::new();
         let group = client
@@ -141,9 +150,12 @@ mod tests {
                 src: AntennaSource::Group,
                 user_list_id: None,
                 user_group_id: Some(group.id),
-                keywords: vec![vec!["kawaii".to_string()], vec!["cute".to_string()]],
+                keywords: Query::from_vec(vec![
+                    vec!["kawaii".to_string()],
+                    vec!["cute".to_string()],
+                ]),
                 #[cfg(feature = "12-19-0")]
-                exclude_keywords: Vec::new(),
+                exclude_keywords: Query::default(),
                 users: Vec::new(),
                 case_sensitive: false,
                 with_replies: true,
@@ -155,7 +167,7 @@ mod tests {
 
     #[tokio::test]
     async fn request_users() {
-        use crate::model::antenna::AntennaSource;
+        use crate::model::{antenna::AntennaSource, query::Query};
 
         let client = TestClient::new();
         let admin = client.admin.me().await;
@@ -168,9 +180,12 @@ mod tests {
                 user_list_id: None,
                 #[cfg(feature = "12-10-0")]
                 user_group_id: None,
-                keywords: vec![vec!["annoucement".to_string()], vec!["notice".to_string()]],
+                keywords: Query::from_vec(vec![
+                    vec!["annoucement".to_string()],
+                    vec!["notice".to_string()],
+                ]),
                 #[cfg(feature = "12-19-0")]
-                exclude_keywords: Vec::new(),
+                exclude_keywords: Query::default(),
                 users: vec![admin.username],
                 case_sensitive: false,
                 with_replies: true,