+++ misskey-api/src/endpoint/antennas/list.rs
@@ -18,7 +18,7 @@ mod tests {
 
     #[tokio::test]
     async fn request() {
-        use crate::model::antenna::AntennaSource;
+        use crate::model::{antenna::AntennaSource, query::Query};
 
         let client = TestClient::new();
         client
@@ -28,9 +28,9 @@ mod tests {
                 user_list_id: None,
                 #[cfg(feature = "12-10-0")]
                 user_group_id: None,
-                keywords: Vec::new(),
+                keywords: Query::default(),
                 #[cfg(feature = "12-19-0")]
-                exclude_keywords: Vec::new(),
+                exclude_keywords: Query::default(),
                 users: Vec::new(),
                 case_sensitive: false,
                 with_replies: false,